Move reverse project parsing to separate class
